Enhances ManagedErrorBus to provide a socket like interface with the class ErrorChannel. It now has `recv_message` and `broadcast_message` using an underlying torch.distributed.store implementation (a distributed k-v store). ErrorChannel serializes and deserializes the messages so that it always receives and returns a Message object.
The manager now has a new `self._error_bus_listening_thread` that listens to the error_bus through the ErrorChannel.

A subtle detail is that when the` _error_bus_listening_thread` puts a message on the message_queue, we will not broadcast it, as we have just received it. We implement this by adding a `sender_type` field into Message.

The ErrorMessage format is augmented to include a 'sender_replica_id'. When processing the message_queue, the error_bus checks whether the sender_replica_id == self.replica_id. If so, it does not broadcast it.

Now, in two instances are messages put into the message_queue:  1. When the `error_bus_listening_thread` receives a message from the ErrorChannel.
2. When the manager calls manager.report_error().

 it puts the error: ErrorMessage into error_bus.message_queue.